[Pull Request Bugfix] Task 8: Send/Confirm Recommendation Mail: Part II #14
