+++ app/code/Bulbulatory/Recommendations/Helper/Config.php
@@ -6,9 +6,12 @@ use Magento\Framework\App\Helper\Context;
 use Magento\Framework\App\Helper\AbstractHelper;
 use Magento\Store\Model\ScopeInterface;
 use Magento\Store\Model\StoreManagerInterface;
+use Psr\Log\LoggerInterface;
 
 class Config extends AbstractHelper
 {
+    const RECOMMENDATION_EMAIL_TEMPLATE  = 'recommendations/general/recommendation_email';
+
     /**
      * Store manager
      *
@@ -17,16 +20,24 @@ class Config extends AbstractHelper
     protected $_storeManager;
 
     /**
+     * @var LoggerInterface
+     */
+    private $logger;
+
+    /**
      * Config constructor.
      * @param Context $context
      * @param StoreManagerInterface $storeManager
+     * @param LoggerInterface $logger
      */
     public function __construct(
         Context $context,
-        StoreManagerInterface $storeManager
+        StoreManagerInterface $storeManager,
+        LoggerInterface $logger
     ) {
         parent::__construct($context);
         $this->_storeManager = $storeManager;
+        $this->logger = $logger;
     }
 
     /**
@@ -42,5 +53,27 @@ class Config extends AbstractHelper
             ScopeInterface::SCOPE_STORE,
             $this->_storeManager->getStore()->getStoreId()
         );
+    }
+
+    /**
+     * @return bool
+     */
+    public function isModuleEnabled(): bool
+    {
+        $recommendationModuleAccess = false;
+        try {
+            $recommendationModuleAccess = (bool) $this->getConfigValue('recommendations/general/enable');
+        }
+        catch (\Throwable $exception)
+        {
+            $this->logger->error($exception->getMessage(), ['exception' => $exception]);
+        }
+
+        return $recommendationModuleAccess;
+    }
+
+    public function getRecommendationMailTemplateId()
+    {
+        return $this->getConfigValue(self::RECOMMENDATION_EMAIL_TEMPLATE);
     }
 }
